updated start and stop scripts

+++ start.sh
@@ -1,6 +1,7 @@
 #! /bin/bash
-
-kill $(pgrep -u $(whoami) node);
+pgrep -o -u $(whoami) node >/dev/null
+code=$?;
+[ $code -eq 0 ] && kill $(pgrep -u $(whoami) node);
 git pull;
 npm i;
 nohup node app.js &
